Add support for notifications on property changes to the LogMessage class
The LogMessage class now implements the INotifyPropertyChanged interface. The PropertyChanged event uses the PropertyChangedEventManager to call an event handler in the context of the thread that registered it. To work properly the registering thread must have a synchronization context that handles marshalling the call into this thread. Furthermore the LogMessage class is thread-safe now.