Fix opening a log file read-only, if file was opened in 'robust' mode last
Write mode 'robust' switches the journaling of the underlying sqlite database to WAL mode. If such a database file is opened read-only, sqlite creates -shm and -wal files beside the database file, if permissions allow that. These files are not removed when the database file is closed. Switching journaling mode to 'delete' when the database file is closed, mitigates the issue.