Add support for populating log files on creation
The LogFile class and the FileBackedLogMessageCollection class now have Create() methods that allow to create a new log file and populate it with an initial message set.
Using the Create() method is much faster than creating an empty file and using AddRange() as database indices are created after populating the collection.
Adding a log message after indices have been created cause indices to be updated as messages are written.