Replace throwing exception on full send queue with boolean return value
Throwing a LogServiceChannelQueueFullException is better from the user perspective, but induces severe performance issues in high load scenarios. The allocation of the exception object alone puts additional load on the garbage collector that freezes threads when cleaning up reducing the overall performance. In addition to the garbage collection overhead excessive exception handling is expensive and therefore a boolean return value indicating whether some data was enqueued successfully is a better fit.