Eliminate bottleneck in LogServiceServer
A LogServiceServerChannel called the LogServiceServer every time it received some data to indicate that it is still active. This involved obtaining a lock to manipulate the channel list. The channel list is kept untouched now. Instead the server thread iterates over the channel list to detect inactive channels that need to be shut down.