Revise access to pipeline stage settings [breaking]
Pipeline stages previously used ProcessingPipelineStageConfigurationBase.GetSetting() to obtain a setting object for named settings. This method had undesirable side effects as it actually registered a setting to allow pipeline stages to use it. The registration added a new setting with a default value. The registration failed, if GetSetting() was called with different default setting values. This was intentional behavior, but confused some people, so we decided to rename GetSetting() to RegisterSetting() as it expresses what it actually does. At the same time we added methods to access settings without registering one with a default value.

RegisterSetting() registers a setting with a specific name and creates a new setting with a default value, if the setting does not exist, yet. The default value does not change an existing setting value.
GetSetting() gets a setting with a specific name and returns null, if it does not exist, yet.
SetSetting() sets a setting with a specific name and creates a new setting with the specified value (but without setting a default value to circumvent clashes with RegisterSettings()).