Revise binding pipeline stage settings [breaking]
Introduced a setting proxy that forwards all setting accesses to the currently bound pipeline stage configuration. If the configuration is exchanged, the proxy rebinds to the new configuration automatically. This avoids breaking the link between a pipeline stage and its configuration. The solution up to now was to invoke the virtual BindSettings() method of the ProcessingPipelineBaseStage class just after the configuration has been exchanged. The first call to this method was done in the constructor of the ProcessingPipelineBaseStage class. This could cause severe issues as the constructor of a derived class has not run at the point the virtual method was called. The override of BindSettings() was working on an incompletely initialized object in this case.

Pipeline stages deriving from the ProcessingPipelineBaseStage class can now call RegisterSetting() to get a setting proxy. BindSettings() has been removed.