Let all log configurations implement IDisposable [breaking]
This enables configuration-agnostic tests to clean up unmanaged resources. Under Linux unit tests targeting the FileBackedLogConfiguration class failed sporadicly, because the user was running out of inotify instances needed by the FileSystemWatcher class. Disposing the watchers should solve the issue. In real-world applications there usually is only one configuration with a single watcher, so not disposing the configuration should not be an issue.