Fix issue with cancellation on shutdown in ElasticsearchPipelineStage
If the ElasticsearchPipelineStage does not complete shutting down within 30 seconds, a cancellation token is signaled to cancel pending send operations. This could lead to an OperationAbortedException to be thrown before the send tasks have actually been set to completed. Disposing these incomplete tasks as part of the cleanup procedure could throw an exception.