Fix broken link between log configurations and log writers and pipeline stages
All configurations now provide a 'Changed' event that is raised when something in the configuration changes. Event handlers are fired using the synchronization context of the thread registering the event, so it is suitable for use in conjunction with GUIs. Raising the 'Changed' event can be suspended temporarily.

The FileBackedLogConfiguration effectively supports reloading its *.gplogconf file on changes now.