Elasticsearch Pipeline Stage: Fix response handling
As the pipeline stage was adjusted to work with data streams the 'index' action was replaced with the 'create' action. The changed action was not properly considered when processing the response. This could lead to sending messages multiple times if at least one operation in a bulk request has failed.